Add ProjectTree module and CLI entry
