+++ mcp_project/__main__.py
@@ -1,0 +1,8 @@
+"""Allow running ``python -m mcp_project`` directly."""
+
+from __future__ import annotations
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
